Validate against dates Chronic can't parse.
Previously if you entered a malformed date the page would just blow
up. Now we store the parsed date but keep the original string in a
temporary attribute (a) so we can determine whether a "nil" parsed
date is because of a malformed string or an empty one, and (b) so we
can present it back to the user if something else goes wrong.

This commit also includes the first appearance of a separate test for
the Publishable module. Previously we would just test its features by
way of the classes that include it.