Don't refresh the page when submitting a page form.
It was jarring when there was an error to go from "comic page with
popup" to "edit-only page." Now errors are presented within the
popup, and if the submission is successful we just replace the
content.

This doesn't work right now if you upload an image, apparently because
of security restrictions. Planning to fix that in the next commit.

"Due to JavaScript security limitations it is impossible to send forms
with (non-empty) file fields via XHR."

http://stackoverflow.com/questions/1740747/jquery-form-plugin-ajax-file-upload-rails