Fixed test bug: "Element not found in the cache".
WebDriver sometimes raised this error with the old code. I think the
element was being removed after we "found" it but before we asked it
if it was visible. For the "close" case we don't really have to worry
about the element being present-but-not-visible, so we can simplify
the code.