Made cancel work after a validation error.
Previously we were binding a click event to the cancel link after the
popup was closed. When we started doing validation inside the popup,
it replaced the cancel link with a new one, with no click binding. Now
we just use .live() to make sure all modal cancel links work the same.